(An Abbreviation of) The LEAD Data Set

These are an abbreviated version of the LEAD Data Set, incorporating variables that I think are most interesting or potentially useful from these data.

Usage

LEAD

Format

A data frame with 3409 observations on the following 12 variables.

obsid

an observational ID from archigos

leveledu

0 = primary, 1 = secondary, 2 = university, 3 = graduate

milservice

did leader have prior military service?

combat

did leader have prior combat experience in military service?

rebel

was leader previously part of a rebel group?

warwin

was leader previously part of a winning war effort as part of military service?

warloss

was leader previously part of a losing war effort as part of military service?

rebelwin

was leader previously part of a winning war effort as part of a rebel group?

rebelloss

was leader previously part of a losing war effort as part of a rebel group?

yrsexper

previous years of experience in politics before becoming a leader

physhealth

does leader have physical health issues?

mentalhealth

does leader have mental health issues?

Details

Data are ported from Ellis et al. (2015). Users who want more of these variables included in peacesciencer should raise an issue on Github.

References

Ellis, Carli Mortenson, Michael C. Horowitz, and Allan C. Stam. 2015. "Introducing the LEAD Data Set." International Interactions 41(4): 718–741.