Add Correlates of War international governmental organizations (IGOs) data to dyad-year or state-year data.

add_igos() allows you to add information from the Correlates of War International Governmental Organizations data to dyad-year or state-year data, matching on Correlates of War system codes.

Usage

add_igos(data)

Arguments

data

a dyad-year data frame (either "directed" or "non-directed") or a state-year data frame.

Value

add_igos() takes a dyad-year data frame or state-year data frame and adds information available from the Correlates of War International Governmental Organizations data. If the data are dyad-year, the function returns the original data with just one additional column for the total number of mutual IGOs for which both members of the dyad are full members. If the data are state-year, the function returns the original data with four additional columns. These are the number of IGOs for which the state is a full member, the number of IGOs for which the state is an associate member, the number of IGOs for which the state is an observer, and the number of IGOs for which the state is involved in any way (i.e. the sum of the other three columns).

Details

The function leans on attributes of the data that are provided by the create_dyadyear() or create_stateyear() function. Make sure that function (or data created by that function) appear at the top of the proverbial pipe.

References

Pevehouse, Jon C.W., Timothy Nordstron, Roseanne W McManus, and Anne Spencer Jamison. 2020. "Tracking Organizations in the World: The Correlates of War IGO Version 3.0 datasets." Journal of Peace Research 57(3): 492-503.

Wallace, Michael, and J. David Singer. 1970. "International Governmental Organization in the Global System, 1815-1964." International Organization 24: 239-87.

Author

Steven V. Miller

Examples

# \donttest{

# just call library(tidyverse) at the top of the pipe
library(magrittr)

cow_ddy %>% add_igos()
#> Joining with `by = join_by(ccode1, ccode2, year)`
#> # A tibble: 2,214,930 × 4
#>    ccode1 ccode2  year dyadigos
#>     <dbl>  <dbl> <dbl>    <dbl>
#>  1      2     20  1920        7
#>  2      2     20  1921        8
#>  3      2     20  1922        8
#>  4      2     20  1923        7
#>  5      2     20  1924        7
#>  6      2     20  1925        8
#>  7      2     20  1926        8
#>  8      2     20  1927        8
#>  9      2     20  1928        8
#> 10      2     20  1929        8
#> # ℹ 2,214,920 more rows

create_stateyears() %>% add_igos()
#> Joining with `by = join_by(ccode, year)`
#> # A tibble: 17,511 × 7
#>    ccode statenme           year sum_igo_full sum_igo_associate sum_igo_observer
#>    <dbl> <chr>             <dbl>        <dbl>             <dbl>            <dbl>
#>  1     2 United States of…  1816            0                 0                0
#>  2     2 United States of…  1817            0                 0                0
#>  3     2 United States of…  1818            0                 0                0
#>  4     2 United States of…  1819            0                 0                0
#>  5     2 United States of…  1820            0                 0                0
#>  6     2 United States of…  1821            0                 0                0
#>  7     2 United States of…  1822            0                 0                0
#>  8     2 United States of…  1823            0                 0                0
#>  9     2 United States of…  1824            0                 0                0
#> 10     2 United States of…  1825            0                 0                0
#> # ℹ 17,501 more rows
#> # ℹ 1 more variable: sum_igo_anytype <dbl>
# }