Add (Select) Leader Experience and Attribute Descriptions (LEAD) Data to Leader-Year or Leader-Dyad-Year Data

add_lead() allows you to add some data recorded in the LEAD data to your leader-year or leader-dyad-year data.

add_lead(data, keep)

Arguments

data
a leader-year or leader-dyad-year data frame

keep
an optional parameter, specified as a character vector, about what leader attributes the user wants to return from this function. If keep is not specified, everything from the LEAD data in this package is returned. Otherwise, the function subsets the LEAD data to just what the user wants.

Value

add_lead() takes a leader-year or leader-dyad-year data frame and adds some data recorded in the LEAD data to it. For leader-dyad-year data, suffices of "1" and "2" are added to the data to indicate attributes of the first leader (obsid1) or the second leader (obsid2), respectively.

References

Ellis, Carli Mortenson, Michael C. Horowitz, and Allan C. Stam. 2015. "Introducing the LEAD Data Set." International Interactions 41(4): 718--741.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)

create_leaderyears() %>% add_lead()

#> Joining, by = c("gwcode", "year")

#> Joining, by = "obsid"

#> # A tibble: 17,686 × 19
#>    obsid     year gwcode leader gender leaderage yrinoffice leadid29 leveledu
#>    <chr>    <dbl>  <dbl> <chr>  <chr>      <dbl>      <int> <chr>       <dbl>
#>  1 USA-1869  1869      2 Grant  M             47          1 A2.9-1          2
#>  2 USA-1869  1870      2 Grant  M             48          2 A2.9-1          2
#>  3 USA-1869  1871      2 Grant  M             49          3 A2.9-1          2
#>  4 USA-1869  1872      2 Grant  M             50          4 A2.9-1          2
#>  5 USA-1869  1873      2 Grant  M             51          5 A2.9-1          2
#>  6 USA-1869  1874      2 Grant  M             52          6 A2.9-1          2
#>  7 USA-1869  1875      2 Grant  M             53          7 A2.9-1          2
#>  8 USA-1869  1876      2 Grant  M             54          8 A2.9-1          2
#>  9 USA-1869  1877      2 Grant  M             55          9 A2.9-1          2
#> 10 USA-1877  1877      2 Hayes  M             55          1 A2.9-4          3
#> # … with 17,676 more rows, and 10 more variables: milservice <dbl>,
#> #   combat <dbl>, rebel <dbl>, warwin <dbl>, warloss <dbl>, rebelwin <dbl>,
#> #   rebelloss <dbl>, yrsexper <dbl>, physhealth <dbl>, mentalhealth <dbl>

create_leaderyears() %>% add_lead(keep = c("yrsexper"))

#> Joining, by = c("gwcode", "year")

#> Joining, by = "obsid"

#> # A tibble: 17,686 × 8
#>    obsid     year gwcode leader gender leaderage yrinoffice yrsexper
#>    <chr>    <dbl>  <dbl> <chr>  <chr>      <dbl>      <int>    <dbl>
#>  1 USA-1869  1869      2 Grant  M             47          1        0
#>  2 USA-1869  1870      2 Grant  M             48          2        0
#>  3 USA-1869  1871      2 Grant  M             49          3        0
#>  4 USA-1869  1872      2 Grant  M             50          4        0
#>  5 USA-1869  1873      2 Grant  M             51          5        0
#>  6 USA-1869  1874      2 Grant  M             52          6        0
#>  7 USA-1869  1875      2 Grant  M             53          7        0
#>  8 USA-1869  1876      2 Grant  M             54          8        0
#>  9 USA-1869  1877      2 Grant  M             55          9        0
#> 10 USA-1877  1877      2 Hayes  M             55          1       12
#> # … with 17,676 more rows

# }