Add Estimates of Leader Willingness to Use Force to Leader-Year Data

add_lwuf() allows you to add estimates of leader willingness to use force to leader-year data or leader-dyad-year data.

Usage

add_lwuf(data, keep)

Arguments

data

a leader-year or leader dyad-year data frame as generated in peacesciencer

keep

an optional argument, specified as a character vector, of the variables from the lwuf data frame the user wants in their data. See the lwuf data and its documentation for more. If the argument is unspecified, the function will return all measures of leader willingness to use force as generated by Carter and Smith.

Value

add_lwuf() takes a leader-year or leader-dyad-year data frame and adds estimates of leader willingness to use force, as generated by Carter and Smith (2020).

Details

See lwuf for more information, but I'll copy-paste it here too.

The letter published by Carter and Smith (2020) contains more information as to what these thetas refer. The "M1" theta is a variation of the standard Rasch model from the boilerplate information in the LEAD data. The authors consider this to be "theoretically relevant" or "risk-related" as these all refer to conflict or risk-taking. The "M2" theta expands on "M1" by including political orientation and psychological characteristics. "M3" and "M4" expand on "M1" and "M2" by considering all 36 variables in the LEAD data.

The authors construct and include all these measures, though their analyses suggest "M2" is the best-performing measure. You should probably consider using theta2_mean as your default estimate of leader willingness to use force in leader-year analyses.

References

Carter, Jeff and Charles E. Smith, Jr. 2020. "A Framework for Measuring Leaders' Willingness to Use Force." American Political Science Review 114(4): 1352–1358.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)

create_leaderyears() %>% add_lwuf()
#> Joining with `by = join_by(obsid)`
#> # A tibble: 17,686 × 15
#>    obsid   leader gwcode gender leaderage  year yrinoffice theta1_mean theta1_sd
#>    <chr>   <chr>   <dbl> <chr>      <dbl> <dbl>      <dbl>       <dbl>     <dbl>
#>  1 USA-18… Grant       2 M             47  1869          1       1.01      0.509
#>  2 USA-18… Grant       2 M             48  1870          2       1.01      0.509
#>  3 USA-18… Grant       2 M             49  1871          3       1.01      0.509
#>  4 USA-18… Grant       2 M             50  1872          4       1.01      0.509
#>  5 USA-18… Grant       2 M             51  1873          5       1.01      0.509
#>  6 USA-18… Grant       2 M             52  1874          6       1.01      0.509
#>  7 USA-18… Grant       2 M             53  1875          7       1.01      0.509
#>  8 USA-18… Grant       2 M             54  1876          8       1.01      0.509
#>  9 USA-18… Grant       2 M             55  1877          9       1.01      0.509
#> 10 USA-18… Hayes       2 M             55  1877          1       0.464     0.545
#> # ℹ 17,676 more rows
#> # ℹ 6 more variables: theta2_mean <dbl>, theta2_sd <dbl>, theta3_mean <dbl>,
#> #   theta3_sd <dbl>, theta4_mean <dbl>, theta4_sd <dbl>
# }