Add rugged terrain information to a data frame

add_rugged_terrain() allows you to add information, however crude, about the "ruggedness" of a state's terrain to your (dyad-year, leader-year, leader-dyad-year, state-year) data.

Usage

add_rugged_terrain(data)

Arguments

data

a data frame with appropriate peacesciencer attributes

Value

add_rugged_terrain() takes a (dyad-year, leader-year, leader-dyad-year, state-year) data frame, whether the primary state identifiers are from the Correlates of War system or the Gleditsch-Ward system, and returns information about the "ruggedness" of the state's terrain. The two indicators returned are the "terrain ruggedness index" calculated by Nunn and Puga (2012) and a logarithmic transformation of how mountainous the state is (as calculated by Gibler and Miller, 2014). The dyad-year (leader-dyad-year) data get four additional columns (i.e. both indicators for both states in the dyad) whereas the state-year data get just the two additional columns.

Details

Please see the information for the underlying data rugged, and the associated R script in the data-raw directory, to see how these data are generated. Importantly, these data are time-agnostic and move slowly. We're talking about geography here. Both data sets benchmark around 1999-2000 and it's a leap of faith to use these data for comparisons across the entirety of the Correlates of War or Gleditsch-Ward system membership. Every use of data of these types have been either cross-sectional snapshots or for making state-to-state comparisons after World War II (think of your prominent civil war studies here). Be mindful about what you expect to get from these data.

The underlying data have both Gleditsch-Ward codes and Correlates of War codes. The merge it makes depends on what you declare as the "master" system at the top of the pipe (e.g.. in create_dyadyears() or create_stateyears()). If, for example, you run create_stateyears(system="cow") and follow it with add_gwcode_to_cow(), the merge will be on the Correlates of War codes and not the Gleditsch-Ward codes. You can see the script mechanics to see how this is achieved.

References

Fearon, James D., and David Laitin, "Ethnicity, Insurgency, and Civil War" American Political Science Review 97: 75–90.

Gibler, Douglas M. and Steven V. Miller. 2014. "External Territorial Threat, State Capacity, and Civil War." Journal of Peace Research 51(5): 634-646.

Nunn, Nathan and Diego Puga. 2012. "Ruggedness: The Blessing of Bad Geography in Africa." Review of Economics and Statistics. 94(1): 20-36.

Riley, Shawn J., Stephen D. DeGloria, and Robert Elliot. 1999. "A Terrain Ruggedness Index That Quantifies Topographic Heterogeneity,” Intermountain Journal of Sciences 5: 23–27.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)

cow_ddy %>% add_rugged_terrain()
#> # A tibble: 2,101,440 × 7
#>    ccode1 ccode2  year rugged1 newlmtnest1 rugged2 newlmtnest2
#>     <dbl>  <dbl> <int>   <dbl>       <dbl>   <dbl>       <dbl>
#>  1      2     20  1920    1.07        3.21   0.775        2.80
#>  2      2     20  1921    1.07        3.21   0.775        2.80
#>  3      2     20  1922    1.07        3.21   0.775        2.80
#>  4      2     20  1923    1.07        3.21   0.775        2.80
#>  5      2     20  1924    1.07        3.21   0.775        2.80
#>  6      2     20  1925    1.07        3.21   0.775        2.80
#>  7      2     20  1926    1.07        3.21   0.775        2.80
#>  8      2     20  1927    1.07        3.21   0.775        2.80
#>  9      2     20  1928    1.07        3.21   0.775        2.80
#> 10      2     20  1929    1.07        3.21   0.775        2.80
#> # … with 2,101,430 more rows

create_stateyears() %>% add_rugged_terrain()
#> Joining, by = "ccode"
#> # A tibble: 16,926 × 5
#>    ccode statenme                  year rugged newlmtnest
#>    <dbl> <chr>                    <int>  <dbl>      <dbl>
#>  1     2 United States of America  1816   1.07       3.21
#>  2     2 United States of America  1817   1.07       3.21
#>  3     2 United States of America  1818   1.07       3.21
#>  4     2 United States of America  1819   1.07       3.21
#>  5     2 United States of America  1820   1.07       3.21
#>  6     2 United States of America  1821   1.07       3.21
#>  7     2 United States of America  1822   1.07       3.21
#>  8     2 United States of America  1823   1.07       3.21
#>  9     2 United States of America  1824   1.07       3.21
#> 10     2 United States of America  1825   1.07       3.21
#> # … with 16,916 more rows

create_stateyears(system = "gw") %>% add_rugged_terrain()
#> Joining, by = "gwcode"
#> # A tibble: 18,463 × 5
#>    gwcode statename                 year rugged newlmtnest
#>     <dbl> <chr>                    <int>  <dbl>      <dbl>
#>  1      2 United States of America  1816   1.07       3.21
#>  2      2 United States of America  1817   1.07       3.21
#>  3      2 United States of America  1818   1.07       3.21
#>  4      2 United States of America  1819   1.07       3.21
#>  5      2 United States of America  1820   1.07       3.21
#>  6      2 United States of America  1821   1.07       3.21
#>  7      2 United States of America  1822   1.07       3.21
#>  8      2 United States of America  1823   1.07       3.21
#>  9      2 United States of America  1824   1.07       3.21
#> 10      2 United States of America  1825   1.07       3.21
#> # … with 18,453 more rows
# }