Archigos: A (Subset of a) Dataset on Political Leaders

These are leader-level data drawn from the Archigos data. Space considerations mean I offer here just a few columns based on these data. Data are version 4.1.

Usage

archigos

Format

A data frame with 3409 observations on the following 11 variables.

gwcode

a numeric vector for the Gleditsch-Ward state code

obsid

a character vector for observation ID

leadid

the unique leader identifier

leader

the leader name

yrborn

the year the leader was born

gender

a categorical variable for leader gender ("M" for men, "W" for women)

startdate

a date for the leader start date

enddate

a date for the leader end date

entry

a character vector for the leader's entry type

exit

a character vector for the leader's exit type

exitcode

a character vector for more information about the leader's exit type

Details

Space considerations mean I can only offer a few columns from the overall data. Archigos data are rich with information. Consult the raw data available on Hein Goeman's website for more.

To best conform with data requirements on CRAN, a few leader names were renamed if they included irregular characters (e.g. umlauts or accents). These leaders, in these particular applications, hav been renamed to "(Juan Orlando) Hernandez" (HON-2014), "(Antonio) Saca Gonzalez" (SAL-2004), "Julian Trujillo Largacha" (COL-1878), "Cesar Gaviria Trujillo" (COL-1990), "Gabriel Garcia Moreno" (ECU-1869), "Marcos A. Morinigo" (PAR-1894-1), "Higinio Morinigo" (PAR-1940), "Sebastian Pinera" (CHL-2010), "Sauli Niinisto" (FIN-2012), "Louis Gerhard De Geer" (SWD-1876), "Stefan Lofven" (SWD-2014), "Lars Lokke Rasmussen" (DEN-2009, DEN-2015), and "Fernando de Araujo" (ETM-2008-1). None of these names contain these special characters in the data here.

For clarity's sake, I renamed the ccode column in the raw data to be gwcode. This is because it may deceive the user peeking into the data that these are not Correlates of War state codes, but Gleditsch-Ward state codes.

References

Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing Archigos: A Dataset of Political Leaders" Journal of Peace Research 46(2): 269–83.