Archigos Yearly Leader Turnover: A Summary

These are yearly summaries of leader turnover from the Archigos data, for use in add_archigos()

Usage

archigossums

Format

A data frame with 14707 observations on the following 7 variables.

gwcode

a numeric vector for the Gleditsch-Ward state code

year

a numeric vector for a referent year

leadertransition

a dummy variable indicating a leader transition in a given year

irregular

a dummy variable indicating an irregular leader transition in a given year

n_leaders

an integer for the number of leaders in a given year

jan1obsid

a character vector for the observation ID of the head of state on Jan. 1 of the referent year

dec31obsid

a character vector for the observation ID of the head of state on Dec. 31 of the referent year

Details

Consult archigos in the same data frame for more information about the data.

References

Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing Archigos: A Dataset of Political Leaders" Journal of Peace Research 46(2): 269–83.