Alliance Treaty Obligations and Provisions (ATOP) Project Data (v. 5.1)

These are directed dyad-year-level data for alliance obligations and provisions from the ATOP project.

Usage

atop_alliance

Format

A data frame with 273,296 observations on the following eight variables.

ccode1

a numeric vector for the Correlates of War state code for the first state

ccode2

a numeric vector for the Correlates of War state code for the second state

year

a numeric vector for the year

atop_defense

a numeric vector that equals 1 if there was an alliance observed with a defense pledge

atop_offense

a numeric vector that equals 1 if there was an alliance observed with a offense pledge

atop_neutral

a numeric vector that equals 1 if there was an alliance observed with a neutrality pledge

atop_nonagg

a numeric vector that equals 1 if there was an alliance observed with a non-aggression pledge

atop_consul

a numeric vector that equals 1 if there was an alliance observed with a consultation pledge

Details

The data-raw directory on the project's Github shows how the data were processed.

References

Leeds, Brett Ashley, Jeffrey M. Ritter, Sara McLaughlin Mitchell, and Andrew G. Long. 2002. "Alliance Treaty Obligations and Provisions, 1815-1944." International Interactions 28: 237-60.