The Minimum Distance Between States in the Correlates of War System, 1886-2019

These are non-directed dyad-year data for the minimum distance between states in the Correlates of War state system from 1886 to 2019. The data are generated from the cshapes package.

Usage

cow_mindist

Format

A data frame with 817053 observations on the following 4 variables.

ccode1

the Correlates of War state system code for the first state

ccode2

the Correlates of War state system code for the second state

year

the year

mindist

the minimum distance between states on Jan. 1 of the year, in kilometers

Details

The data are generated from the cshapes package. Data are automatically generated (by default) as directed dyad-years. I elect to make them non-directed for space considerations. Making non-directed dyad-year data into directed dyad-year data isn't too difficult in R. It just looks weird to see the code that does it.

Previous versions of these data were for the minimum distance as of Dec. 31 of the referent year. These are now Jan. 1. Most of the data I provide elsewhere in this package are to be understood as the data as they were at the start of the year. add_minimum_distance() permits greater flexibility with this option, but only for the remote and augmented version of the data. Check the documentation of that function for more.

References

Schvitz, Guy, Luc Girardin, Seraina Ruegger, Nils B. Weidmann, Lars-Erik Cederman,and Kristian Skrede Gleditsch. 2022. "Mapping The International System, 1886-2017: The CShapes 2.0 Dataset." Journal of Conflict Resolution. 66(1): 144-161.

Weidmann, Nils B. and Kristian Skrede Gleditsch. 2010. "Mapping and Measuring Country Shapes: The cshapes Package." The R Journal 2(1): 18-24