Correlates of War State System Membership Data (1816-2016)

These are the Correlates of War state system membership data.

Usage

cow_states

Format

A data frame with 243 observations on the following 10 variables.

stateabb

a character vector for the state abbreviation

ccode

a numeric vector for the Correlates of War country code

statenme

a character vector for the state name

styear

the start year in the system

stmonth

the start month in the system

stday

the start day in the system

endyear

the end year in the system

endmonth

the end month in the system

endday

the end day in the system

version

a version identifier

Details

Data are provided "as-is" with no additional re-cleaning before inclusion into this data set.

The functions that previously used these data no longer use these data. They instead use a copy of the data in the isard package I also maintain.

References

Correlates of War Project. 2017. "State System Membership List, v2016." Online, https://correlatesofwar.org/data-sets/state-system-membership/