Directed Dyadic Dispute-Year Data with No Duplicate Dyad-Years (GML, v. 2.2.1)

These are directed dyadic dispute year data derived from the Gibler-Miller-Little (GML) Militarized Interstate Dispute (MID) project. Data are from version 2.2.1. These were whittled to where there is no duplicate dyad-years. Its primary aim here is merging into a dyad-year data frame.

Usage

gml_mid_ddydisps

Format

A data frame with 9262 observations on the following 25 variables.

dispnum

a numeric vector for the dispute number

ccode1

a numeric vector for the focal state in the dyad

ccode2

a numeric vector for the target state in the dyad

year

a numeric vector for the dispute-year

gmlmidongoing

a numeric vector for whether there was a dispute ongoing in that year

gmlmidonset

a numeric vector for whether it was the onset of a new dispute (or new participant-entry into a recurring dispute)

sidea1

is ccode1 on side A of the dispute?

sidea2

is ccode2 on side A of the dispute?

fatality1

a numeric vector for the overall fatality level of ccode1 in the dispute

fatality2

a numeric vector for the overall fatality level of ccode2 in the dispute

fatalpre1

a numeric vector for the known fatalities (with precision) for ccode1 in the dispute

fatalpre2

a numeric vector for the known fatalities (with precision) for ccode2 in the dispute

hiact1

a numeric vector for the highest action of ccode1 in the dispute

hiact2

a numeric vector for the highest action of ccode2 in the dispute

hostlev1

a numeric vector for the hostility level of ccode1 in the dispute

hostlev2

a numeric vector for the hostility level of ccode2 in the dispute

orig1

is ccode1 an originator of the dispute?

orig2

is ccode2 an originator of the dispute?

fatality

a numeric vector for the fatality level of the dispute

hostlev

a numeric vector for the hostility level of the MID

mindur

a numeric vector for the minimum duration of the MID

maxdur

a numeric vector for the maximum duration of the MID

recip

a numeric vector for whether a MID was reciprocated

stmon

a numeric vector for the start month of the MID

Details

The process of creating these is described at one of the references below. Importantly, these data are somewhat "naive." That is: they won't tell you, for example, that Brazil and Japan never directly fought each other during World War II. Instead, it will tell you that there were two years of overlap for the two on different sides of the conflict and that the highest action for both was a war. The data are thus similar to what the EUGene program would create for users back in the day. Use these data with that limitation in mind.

References

Miller, Steven V. 2021. "How to (Meticulously) Convert Participant-Level Dispute Data to Dyadic Dispute-Year Data in R." URL: http://svmiller.com/blog/2021/05/convert-cow-mid-data-to-dispute-year/

Gibler, Douglas M., Steven V. Miller, and Erin K. Little. 2016. “An Analysis of the Militarized Interstate Dispute (MID) Dataset, 1816-2001.” International Studies Quarterly 60(4): 719-730.