Conventional Arms Races During Periods of Rivalry

This is a simple data set of 71 arms races reported by Gibler et al. in their 2005 article in Journal of Peace Research.

Usage

grh_arms_races

Format

A data frame the following five variables.

race_id

the arms race identifier

ccode1

a numeric vector for the Correlates of War state code for the first state

ccode2

a numeric vector for the Correlates of War state code for the second state

styear

the start year for the arms race

endyear

the end year for the arms race

Details

Data are taken from the appendix of Gibler, Rider, and Hutchison's 2005 article in Journal of Peace Research. Read the article and appendix for more information about coding procedures.

References

Gibler, Douglas M., Toby J. Rider, and Marc L. Hutchison. 2005. "Taking Arms Against a Sea of Troubles: Conventional Arms Races during Periods of Rivalry" Journal of Peace Research 42(2): 131–47.