Gleditsch-Ward (Independent States) System Membership Data (1816-2017)

These are the independent states in Gleditsch and Ward's data.

gw_states

Format

A data frame with 216 observations on the following 5 variables.

gwcode

a numeric vector for the Gleditsch-Ward country code

stateabb

a character vector for state abbreviation

statename

a character vector for the state name

startdate

the start date in the data

enddate

the end date in the data

Details

Data originally provided by Gleditsch with no column names. Column names were added before some light re-cleaning in order to generate these data.

References

Gleditsch, Kristian S. and Michael D. Ward. 1999. "A Revised List of Independent States since the Congress of Vienna" International Interactions 25(4): 393--413.