A BibTeX Data Frame of Citations

This is a BibTeX file, loaded as a data frame, to assist the user in properly citing the source material that is used in this package.

ps_bib

Format

A data frame with the following columns.

CATEGORY

the BibTeX entry type

BIBTEXKEY

the BibTeX unique entry key

ADDRESS

another BibTeX field

ANNOTE

another BibTeX field

AUTHOR

a list of authors for this entry

BOOKTITLE

another BibTeX field, for book title (if appropriate)

CHAPTER

another BibTeX field, for chapter (if appropriate)

CROSSREF

another BibTeX field

EDITION

another BibTeX field, for edition of book (if appropriate)

EDITOR

another BibTeX field, for book editor (if appropriate)

HOWPUBLISHED

another BibTeX field

INSTITUTION

another BibTeX field

JOURNAL

another BibTeX field, for the journal name (if appropriate)

KEY

another BibTeX field

MONTH

another BibTeX field

NOTE

another BibTeX field

NUMBER

another BibTeX field, for journal volume number (if appropriate)

ORGANIZATION

another BibTeX field

PAGES

another BibTeX field, for pages of the entry

PUBLISHER

another BibTeX field, for book publisher (if appropriate)

SCHOOL

another BibTeX field

SERIES

another BibTeX field

TITLE

another BibTeX field, for title of the entry

TYPE

another BibTeX field

VOLUME

another BibTeX field, for journal volume (if appropriate)

YEAR

another BibTeX field, for year of publication

KEYWORDS

another BibTeX field, used primarily for selective filtering in this package

URL

another BibTeX field, for website (if appropriate)

OWNER

another BibTeX field

TIMESTAMP

another BibTeX field, used occasionally when I started populating my master file (you will see some old entries here)

DOI

another BibTeX field, for a digital object identifier (used rarely)

EPRINT

another BibTeX field

JOURNALTITLE

another BibTeX field, which I think is actually a BibLaTeX field

ISSN

another BibTeX field

ABSTRACT

another BibTeX field, for entry abstract (if appropriate)

DATE.ADDED

another BibTeX field

DATE.MODIFIED

another BibTeX field

Details

See data-raw directory for how these data were generated. The data were created by bib2df, which is now a package dependency. I assume the user has some familiarity with BibTeX. Some entries were copy-pasted from my master bibliography file that I started in 2008 or so.