Get BibTeX Entries Associated with peacesciencer Data and Functions

ps_cite() allows the user to get citations to scholarship that they should include in their papers that incorporate the functions and data in this package.

Usage

ps_cite(x, column = "keywords")

Arguments

x

a character vector

column

a character vector for the particular column of ps_bib the user wants to search. The default here is "keywords", which searches the KEYWORDS column in ps_bib for the most general search. The other option is "bibtexkey", which will search the BIBTEXKEY column in ps_bib. Use the latter option more for pairing with output from ps_version()

Value

ps_cite() takes a character vector and scans the ps_bib data in this package to return a BibTeX citation (or citations) for the researcher to use to properly cite the material they are getting from this package. The citations are returned as a full BibTeX entry (or entries) that they can copy-paste into their own BibTeX file.

Details

The base functionality here is simple pattern-matching on keywords in ps_bib. This simple pattern-matching is in base R. I assume the user has some familiarity with BibTeX.

Author

Steven V. Miller

Examples

# Cite the package
ps_cite("peacesciencer")
#> @ARTICLE{peacesciencer-package,
#>   AUTHOR = {Steven V. Miller},
#>   JOURNAL = {Conflict Management and Peace Science},
#>   TITLE = {peacesciencer}: An R Package for Quantitative Peace Science Research},
#>   YEAR = {2022},
#>   KEYWORDS = {peacesciencer, add_capital_distance(), add_ccode_to_gw(), add_gwcode_to_cow(), capitals},
#>   URL = {http://svmiller.com/peacesciencer/}} 
#>