The Version Numbers for Data Included in peacesciencer

This is a simple data set that communicates the version numbers of data included in this package. It's a companion to the data frame ps_bib, and other information functions like ps_cite() and ps_version(). The latter uses this data set.

Usage

ps_data_version

Format

A data frame the following four variables.

category

a category for the type of data

data

the name of the particular data source coinciding with the category

version

the version number included in peacesciencer for this data source

bibtexkey

a character key for the BibTeX key corresponding with an appropriate citation in ps_bib

Details

Version numbers that are years should be understood as data sources with no formal version numbering system, per se. Instead, they communicate a year of last update. For example, the Correlates of War does not formally version number its state system data as it does its MID data. Likewise, the Anders et al. (2020) simulations of population and surplus/gross domestic product are not formally versioned, per se. Instead, the data were published and last updated in 2020.