Thompson and Dreyer's (2012) Strategic Rivalries, 1494-2010

A simple summary of all strategic (inter-state) rivalries from Thompson and Dreyer (2012).

td_rivalries

Format

A data frame with 197 observations on the following 10 variables.

rivalryno

a numeric vector for the rivalry number

rivalryname

a character vector for the rivalry name

ccode1

the Correlates of War state code for the state with the lowest Correlates of War state code in the rivalry

ccode2

the Correlates of War state code for the state with the highest Correlates of War state code in the rivalry

styear

a numeric vector for the start year of the rivalry

endyear

a numeric vector for the end year of the rivalry

region

a character vector for the region of the rivalry, per Thompson and Dreyer (2012)

type1

a character vector for the primary type of the rivalry (spatial, positional, ideological, or interventionary)

type2

a character vector for the secondary type of the rivalry, if applicable (spatial, positional, ideological, or interventionary)

type3

a character vector for the tertiary type of the rivalry, if applicable (spatial, positional, ideological, or interventionary)

Details

Information gathered from the appendix of Thompson and Dreyer (2012). Ongoing rivalries are right-bound at 2010, the date of publication for Thompson and Dreyer's handbook. Users are free to change this if they like. Data are effectively identical to strategic_rivalries in stevemisc, but include some behind-the-scenes processing (described in a blog post on http://svmiller.com) that is available to see on the project's Github repository. The data object is also renamed to avoid a conflict.

References

Miller, Steven V. 2019. "Create and Extend Strategic (International) Rivalry Data in R". URL: http://svmiller.com/blog/2019/10/create-extend-strategic-rivalry-data-r/

Thompson, William R. and David Dreyer. 2012. Handbook of International Rivalries. CQ Press.