Estimates from a Random Item Response Model of External Territorial Threat, 1816-2010

This is a state-year data set on (latent) estimates of external territorial threat. Data correspond with a publication in Journal of Global Security Studies.

Usage

terrthreat

Format

A data frame with 14781 observations on the following 10 variables.

ccode

a Correlates of War state code

year

a year

lterrthreat

an estimate of latent external territorial threat for the state in a given year

sd

the standard deviation of simulated, latent external territorial threat

lwr

a lower bound estimate of simulated, latent external territorial threat

upr

an upper bound estimate of simulated, latent external territorial threat

m_lterrthreat

another estimate of latent external territorial threat for the state in a given year

m_sd

another standard deviation of simulated, latent external territorial threat

m_lwr

another lower bound estimate of simulated, latent external territorial threat

m_upr

another upper bound estimate of simulated, latent external territorial threat

Details

The variables with the prefix of m_ communicate alternate estimates in which the state-year-level estimate of territorial threat derived from dyadic data is weighted by the minimum distance between pairs of states. The pertinent variables without this prefix communicate what I (the author!) treat as the standard measure of latent, external territorial threat in which the estimates derived from the dyadic data are weighted by capital distance. You can see the clear corollaries to other functions and data in this package, like the kind used in add_minimum_distance() and add_capital_distance().

The lower and upper bounds communicate 90% intervals.

References

Miller, Steven V. 2022. "A Random Item Response Model of External Territorial Threat, 1816-2010" Journal of Global Security Studies 7(4): ogac012.