Add Correlates of War trade data to a dyad-year or state-year data frame

add_cow_trade() allows you to add Correlates of War alliance data to a dyad-year data frame

add_cow_trade(data)

Arguments

data
a dyad-year data frame (either "directed" or "non-directed") or a state-year data frame

Value

add_cow_trade() takes a dyad-year data frame or state-year data frame and adds information about the volume of trade in that given dyad-year or state-year. For the state-year data, these are minimally the sum of all imports and the sum of all exports. For dyad-year data, this function returns the value of imports in current million USD in the first country from the second country (and vice-versa) along with their "smooth" equivalents.

Details

For the dyad-year data, there must be some kind of information loss in order to work within the limited space available to this package. This package loads a truncated version of the data from my website. It will also load these data every time you use the function for dyad-year data. This implies 1) you probably should not use this function unless you earnestly want these data, 2) this function won't work for you without an active internet connection, and 3) this will be one of the slowest functions in the entire package.

References

Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009. "TRADING DATA: Evaluating our Assumptions and Coding Rules." Conflict Management and Peace Science. 26(5): 471-491.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)

# assumes an active internet connection
cow_ddy %>% add_cow_trade()

#> Joining, by = c("ccode1", "ccode2", "year")

#> # A tibble: 2,063,670 x 7
#>    ccode1 ccode2  year flow2 flow1 smoothflow2 smoothflow1
#>     <dbl>  <dbl> <dbl> <dbl> <dbl>       <dbl>       <dbl>
#>  1      2     20  1920  735.  612.        735.        612.
#>  2      2     20  1921  443.  335.        443.        335.
#>  3      2     20  1922  503.  364.        503.        364.
#>  4      2     20  1923  598.  416         598.        416 
#>  5      2     20  1924  496.  399.        496.        399.
#>  6      2     20  1925  608.  474.        608.        474.
#>  7      2     20  1926  610.  975.        610.        474.
#>  8      2     20  1927  718.  475.        718.        475.
#>  9      2     20  1928  866.  489.        866.        489.
#> 10      2     20  1929  846.  503         846.        503 
#> # … with 2,063,660 more rows

# }