Add Estimates of Leader Willingness to Use Force to Leader-Year Data

add_lwuf() allows you to add estimates of leader willingness to use force to leader-year data or leader-dyad-year data.

add_lwuf(data, keep)

Arguments

data
a leader-year or leader dyad-year data frame as generated in peacesciencer

keep
an optional argument, specified as a character vector, of the variables from the lwuf data frame the user wants in their data. See the lwuf data and its documentation for more. If the argument is unspecified, the function will return all measures of leader willingness to use force as generated by Carter and Smith.

Value

add_lwuf() takes a leader-year or leader-dyad-year data frame and adds estimates of leader willingness to use force, as generated by Carter and Smith (2020).

Details

See lwuf for more information, but I'll copy-paste it here too.

The letter published by Carter and Smith (2020) contains more information as to what these thetas refer. The "M1" theta is a variation of the standard Rasch model from the boilerplate information in the LEAD data. The authors consider this to be "theoretically relevant" or "risk-related" as these all refer to conflict or risk-taking. The "M2" theta expands on "M1" by including political orientation and psychological characteristics. "M3" and "M4" expand on "M1" and "M2" by considering all 36 variables in the LEAD data.

The authors construct and include all these measures, though their analyses suggest "M2" is the best-performing measure. You should probably consider using theta2_mean as your default estimate of leader willingness to use force in leader-year analyses.

References

Carter, Jeff and Charles E. Smith, Jr. 2020. "A Framework for Measuring Leaders' Willingness to Use Force." American Political Science Review 114(4): 1352--1358.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)

create_leaderyears() %>% add_lwuf()

#> Joining, by = c("gwcode", "year")

#> Joining, by = "obsid"

#> # A tibble: 17,686 × 16
#>    obsid     year gwcode leader gender leaderage yrinoffice leadid29 theta1_mean
#>    <chr>    <dbl>  <dbl> <chr>  <chr>      <dbl>      <int> <chr>          <dbl>
#>  1 USA-1869  1869      2 Grant  M             47          1 A2.9-1         1.01 
#>  2 USA-1869  1870      2 Grant  M             48          2 A2.9-1         1.01 
#>  3 USA-1869  1871      2 Grant  M             49          3 A2.9-1         1.01 
#>  4 USA-1869  1872      2 Grant  M             50          4 A2.9-1         1.01 
#>  5 USA-1869  1873      2 Grant  M             51          5 A2.9-1         1.01 
#>  6 USA-1869  1874      2 Grant  M             52          6 A2.9-1         1.01 
#>  7 USA-1869  1875      2 Grant  M             53          7 A2.9-1         1.01 
#>  8 USA-1869  1876      2 Grant  M             54          8 A2.9-1         1.01 
#>  9 USA-1869  1877      2 Grant  M             55          9 A2.9-1         1.01 
#> 10 USA-1877  1877      2 Hayes  M             55          1 A2.9-4         0.464
#> # … with 17,676 more rows, and 7 more variables: theta1_sd <dbl>,
#> #   theta2_mean <dbl>, theta2_sd <dbl>, theta3_mean <dbl>, theta3_sd <dbl>,
#> #   theta4_mean <dbl>, theta4_sd <dbl>

# }