Add Correlates of War National Military Capabilities Data

add_nmc() allows you to add the Correlates of War National Material Capabilities data to dyad-year or state-year data.

add_nmc(data)

Arguments

data
a dyad-year data frame (either "directed" or "non-directed") or a state-year data frame.

Value

add_nmc() takes a dyad-year data frame or state-year data frame and adds information about the national material capabilities for the state or two states in the dyad in a given year. If the data are dyad-year, the function adds 12 total columns for the first state (i.e. ccode1) and the second state (i.e. ccode2) for all estimates of national military capabilities provided by the Correlates of War project. If the data are state-year, the function returns six additional columns to the original data that contain that same information for a given state in a given year.

Details

The function leans on attributes of the data that are provided by the create_dyadyear() or create_stateyear() function. Make sure that function (or data created by that function) appear at the top of the proverbial pipe.

References

Singer, J. David, Stuart Bremer, and John Stuckey. (1972). "Capability Distribution, Uncertainty, and Major Power War, 1820-1965." in Bruce Russett (ed) Peace, War, and Numbers, Beverly Hills: Sage, 19-48.

Singer, J. David. 1987. "Reconstructing the Correlates of War Dataset on Material Capabilities of States, 1816-1985" International Interactions, 14: 115-32.

Author

Steven V. Miller

Examples

# just call `library(tidyverse)` at the top of the your script
library(magrittr)

cow_ddy %>% add_nmc()

#> # A tibble: 2,063,670 × 17
#>    ccode1 ccode2  year  milex1 milper1 irst1   pec1  tpop1 upop1 cinc1 milex2
#>     <dbl>  <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>
#>  1      2     20  1920 1657118     343 42809 743808 106461 27428 0.290  10755
#>  2      2     20  1921 1116342     387 20101 622541 108538 28210 0.253  10209
#>  3      2     20  1922  860853     270 36173 641311 110049 29013 0.256  10028
#>  4      2     20  1923  678256     247 45665 834889 111947 29840 0.272  13316
#>  5      2     20  1924  570142     261 38540 762070 114109 30690 0.254  12824
#>  6      2     20  1925  589706     252 46122 790029 115829 31565 0.254  12984
#>  7      2     20  1926  558004     247 49069 852304 117397 32464 0.263  13936
#>  8      2     20  1927  596501     249 45656 842978 119035 33389 0.239  16745
#>  9      2     20  1928  678100     251 52371 833446 120509 34340 0.240  18862
#> 10      2     20  1929  701300     255 57339 903141 121767 35318 0.240  21058
#> # … with 2,063,660 more rows, and 6 more variables: milper2 <dbl>, irst2 <dbl>,
#> #   pec2 <dbl>, tpop2 <dbl>, upop2 <dbl>, cinc2 <dbl>

create_stateyears() %>% add_nmc()

#> Joining, by = c("ccode", "year")

#> # A tibble: 16,731 × 10
#>    ccode statenme               year milex milper  irst   pec  tpop  upop   cinc
#>    <dbl> <chr>                 <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1     2 United States of Ame…  1816  3823     17    80   254  8659   101 0.0397
#>  2     2 United States of Ame…  1817  2466     15    80   277  8899   106 0.0358
#>  3     2 United States of Ame…  1818  1910     14    90   302  9139   112 0.0361
#>  4     2 United States of Ame…  1819  2301     13    90   293  9379   118 0.0371
#>  5     2 United States of Ame…  1820  1556     15   110   303  9618   124 0.0371
#>  6     2 United States of Ame…  1821  1612     11   100   321  9939   130 0.0342
#>  7     2 United States of Ame…  1822  1079     10   100   332 10268   136 0.0329
#>  8     2 United States of Ame…  1823  1170     11   110   345 10596   143 0.0331
#>  9     2 United States of Ame…  1824  1261     11   110   390 10924   151 0.0330
#> 10     2 United States of Ame…  1825  1336     11   120   424 11252   158 0.0342
#> # … with 16,721 more rows