Add Peace Years to Your Conflict Data

add_peace_years() calculates peace years for your ongoing dyadic conflicts. The function works for both the CoW-MID data and the Gibler-Miller-Little (GML) MID data.

add_peace_years(data, pad = FALSE)

Arguments

data
a dyad-year data frame (either "directed" or "non-directed")

pad
an optional parameter, defaults to FALSE. If TRUE, the peace-year calculations fill in cases where panels are unbalanced/have gaps. Think of a state like Germany disappearing for 45 years as illustrative of this.

If keep is not specified in the function, the ensuing output returns everything.

Value

add_peace_years() takes a dyad-year data frame and adds peace years for ongoing dyadic conflicts.

Details

The function internally uses sbtscs() from stevemisc. In the interest of full disclosure, sbtscs() leans heavily on btscs() from DAMisc. I optimized some code for performance.

Importantly, the underlying function (sbtscs() in stevemisc, by way of btscs() in DAMisc) has important performance issues if you're trying to run it when your event data are sandwiched by observations without any event data. Here's what I mean. Assume you got the full Gleditsch-Ward state-year data from 1816 to 2020 and then added the UCDP armed conflict data to it. If you want the peace-years for this, the function will fail because every year from 1816 to 1945 (along with 2020, as of writing) have no event data. You can force the function to "not fail" by setting pad = TRUE as an argument, but it's not clear this is advisable for this reason. Assume you wanted event data in UCDP for just the extrasystemic onsets. The data start in 1946 and, in 1946, the United Kingdom, Netherlands, and France had extrasystemic conflicts. For *all* years before 1946, the events are imputed as 1 for those countries that had 1s in the first year of observation and everyone else is NA and implicitly assumed to be a zero. For those NAs, the function runs a sequence resulting in some wonky spells in 1946 that are not implied by (the absence of) the data. In fact, none of those are implied by the asbsence of data before 1946.

The function works just fine if you truncate your temporal domain to reflect the nature of your event data. Basically, if you want to use this function more generally, filter your dyad-year or state-year data to make sure there are no years without any event data recorded (e.g. why would you have a CoW-MID analyses of dyad-years with observations before 1816?). This is less a problem when years with all-NAs succeed (and do not precede) the event data. For example, the UCDP conflict data run from 1946 to 2019 (as of writing). Having 2020 observations in there won't compromise the function output when pad = TRUE is included as an argument.

References

Armstrong, Dave. 2016. ``DAMisc: Dave Armstrong's Miscellaneous Functions.'' R package version 1.4-3.

Miller, Steven V. 2017. ``Quickly Create Peace Years for BTSCS Models with sbtscs in stevemisc.'' http://svmiller.com/blog/2017/06/quickly-create-peace-years-for-btscs-models-with-stevemisc/

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)
cow_ddy %>%
add_gml_mids(keep = NULL) %>%
add_cow_mids(keep = NULL) %>%
add_contiguity() %>%
add_cow_majors() %>%
filter_prd()  %>%
add_peace_years()

#> Joining, by = c("ccode1", "ccode2", "year")

#> Joining, by = c("ccode1", "ccode2", "year")

#> Joining, by = c("ccode1", "ccode2", "year")

#> Joining, by = c("year", "dyad")

#> Joining, by = c("year", "dyad")

#> # A tibble: 246,314 × 13
#>    ccode1 ccode2  year gmlmidonset gmlmidongoing cowmidonset cowmidongoing
#>     <dbl>  <dbl> <dbl>       <dbl>         <dbl>       <dbl>         <dbl>
#>  1      2     20  1920           0             0           0             0
#>  2      2     20  1921           0             0           0             0
#>  3      2     20  1922           0             0           0             0
#>  4      2     20  1923           0             0           0             0
#>  5      2     20  1924           0             0           0             0
#>  6      2     20  1925           0             0           0             0
#>  7      2     20  1926           0             0           0             0
#>  8      2     20  1927           0             0           0             0
#>  9      2     20  1928           0             0           0             0
#> 10      2     20  1929           0             0           0             0
#> # … with 246,304 more rows, and 6 more variables: conttype <dbl>,
#> #   cowmaj1 <dbl>, cowmaj2 <dbl>, prd <dbl>, cowmidspell <dbl>,
#> #   gmlmidspell <dbl>

# }