Add Thompson et al. (2021) strategic rivalry data to state-year or dyad-year data frame

add_strategic_rivalries() merges in Thompson et al. (2021) strategic rivalry data to a dyad-year or state-year data frame. The right-bound, as of right now, are bound at 2020.

Usage

add_strategic_rivalries(data)

Arguments

data

a dyad-year data frame (either "directed" or "non-directed")

Value

add_strategic_rivalries() takes a state-year or dyad-year data frame and adds information about ongoing strategic rivalries. It will also include a simple dummy variable for whether there was an ongoing rivalry in the year or not in the dyad-year data. For state-year data, it returns the count of ongoing strategic rivalries for the state in the year meeting a certain criteria (i.e. whether the state has an interventionary, ideological, positional, or spatial rivalry in an ongoing year, and how many).

Details

add_strategic_rivalries() will include some other information derived from the rivalry data that the user may not want (e.g. start year of the rivalry). Feel free to select those out after the fact.

Underneath the hood, the function subsets data to just all rivalry-year observations on or after 1816. This will be in place as long as the Correlates of War state system has a left-bound of 1816 on its temporal domain.

This function includes an on-the-fly adjustment for the Austria-Serbia rivalry (tssr_id = 76). In this case, the last two years of that rivalry are afforded to Austria (ccode = 305) when the bulk of the rivalry pertained to the larger Austria-Hungary (ccode = 300). Previous versions of this function that used the Thompson and Dreyer (2012) strategic rivalry data did the same thing. It was rivalry #79 in that case.

I could technically make such an adjustment on the fly for the France-Germany rivalry as well in these data (tssr_id = 22). If the rivalry concludes in 1955, per the data, it's conceivable that this rivalry should apply to the first two years of statehood for West/East Germany. However, I lean on an earlier version of the data in which this rivalry was classified as a European great power rivalry (see: rivalryno = 22 in td_rivalries). Thus, it makes sense to square the actual rivalry end date with Germany's time as a great power (and its elimination from the international system following the second world war).

I elect to not support the information on principal and asymmetric principal rivalries for the time being. This is subject to change in future versions of the package.

References

Miller, Steven V. 2019. "Create and Extend Strategic (International) Rivalry Data in R". URL: https://svmiller.com/blog/2019/10/create-extend-strategic-rivalry-data-r/

Thompson, William R., Kentaro Sakuwa, and Prashant Hosur Suhas. 2021. Analyzing Strategic Rivalries in World Politics: Types of Rivalry, Regional Variation, and Escalation/De-escalation. Springer.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)
cow_ddy %>% add_strategic_rivalries()
#> Joining with `by = join_by(ccode1, ccode2, year)`
#> # A tibble: 2,214,930 × 9
#>    ccode1 ccode2  year tssr_id positional spatial ideological interventionary
#>     <dbl>  <dbl> <int>   <int>      <dbl>   <dbl>       <dbl>           <dbl>
#>  1      2     20  1920      NA          0       0           0               0
#>  2      2     20  1921      NA          0       0           0               0
#>  3      2     20  1922      NA          0       0           0               0
#>  4      2     20  1923      NA          0       0           0               0
#>  5      2     20  1924      NA          0       0           0               0
#>  6      2     20  1925      NA          0       0           0               0
#>  7      2     20  1926      NA          0       0           0               0
#>  8      2     20  1927      NA          0       0           0               0
#>  9      2     20  1928      NA          0       0           0               0
#> 10      2     20  1929      NA          0       0           0               0
#> # ℹ 2,214,920 more rows
#> # ℹ 1 more variable: ongoingrivalry <dbl>

create_stateyears() %>% add_strategic_rivalries()
#> # A tibble: 17,511 × 7
#>    ccode cw_name     year n_ideological n_interventionary n_positional n_spatial
#>    <dbl> <chr>      <int>         <dbl>             <dbl>        <dbl>     <dbl>
#>  1     2 United St…  1816             0                 0            1         2
#>  2     2 United St…  1817             0                 0            1         2
#>  3     2 United St…  1818             0                 0            1         2
#>  4     2 United St…  1819             0                 0            1         2
#>  5     2 United St…  1820             0                 0            1         1
#>  6     2 United St…  1821             0                 0            2         2
#>  7     2 United St…  1822             0                 0            2         2
#>  8     2 United St…  1823             0                 0            2         2
#>  9     2 United St…  1824             0                 0            2         2
#> 10     2 United St…  1825             0                 0            2         2
#> # ℹ 17,501 more rows
# }