Add Thompson and Dreyer's (2012) strategic rivalry data to dyad-year data frame

add_strategic_rivalries() merges in Thompson and Dreyer's (2012) strategic rivalry data to a dyad-year data frame. The right-bound, as of right now, are bound at 2010.

add_strategic_rivalries(data, across_types = 1)

Arguments

data
a dyad-year data frame (either "directed" or "non-directed")

across_types
optional, relevant for state-year, takes a value of 1, 2, or 3 to look for whether one of three types fits criteria for ideological, interventionary, positional, spatial rivalry. Defaults to 1.

Value

add_strategic_rivalries() takes a dyad-year data frame and adds information about ongoing strategic rivalries. It will also include a simple dummy variable for whether there was an ongoing rivalry in the year or not. For state-year data, it returns the count of ongoing strategic rivalries for the state in the year meeting a certain criteria (i.e. whether the state has an interventionary, ideological, positional, or spatial rivalry in an ongoing year, and how many).

Details

add_strategic_rivalries() will include some other information derived from the rivalry data that the user may not want (e.g. start year of the rivalry). Feel free to select those out after the fact. Function includes an on-the-fly adjustment for Austria for rivalry #79. In this case, the Austria-Serbia rivalry continues for two years after Austria-Hungary (ccode: 300) became Austria (ccode: 305).

The across_types argument is optional and observed for only state-year calls. It defaults to 1. At the default, the function looks into the rivalry data (in td_rivalries) and focuses on the `type1` column. If, say, a state has an ongoing rivalry and it is primarily spatial, it codes that as a spatial rivalry. Assume you input `across_types = 2`, the function then looks across both the `type1` and `type2` columns to see if there is a spatial component to the rivalry as either its primary or secondary dimension. If so, it codes that as a 1. across_types must be 1, 2, or 3.

References

Miller, Steven V. 2019. "Create and Extend Strategic (International) Rivalry Data in R". URL: http://svmiller.com/blog/2019/10/create-extend-strategic-rivalry-data-r/

Thompson, William R. and David Dreyer. 2012. Handbook of International Rivalries. CQ Press.

Author

Steven V. Miller

Examples

# \donttest{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)
cow_ddy %>% add_strategic_rivalries()

#> Joining, by = c("ccode1", "ccode2", "year")

#> # A tibble: 2,063,670 x 12
#>    ccode1 ccode2  year rivalryno rivalryname styear endyear region type1 type2
#>     <dbl>  <dbl> <int>     <dbl> <chr>        <dbl>   <dbl> <chr>  <chr> <chr>
#>  1      2     20  1920        NA NA              NA      NA NA     NA    NA   
#>  2      2     20  1921        NA NA              NA      NA NA     NA    NA   
#>  3      2     20  1922        NA NA              NA      NA NA     NA    NA   
#>  4      2     20  1923        NA NA              NA      NA NA     NA    NA   
#>  5      2     20  1924        NA NA              NA      NA NA     NA    NA   
#>  6      2     20  1925        NA NA              NA      NA NA     NA    NA   
#>  7      2     20  1926        NA NA              NA      NA NA     NA    NA   
#>  8      2     20  1927        NA NA              NA      NA NA     NA    NA   
#>  9      2     20  1928        NA NA              NA      NA NA     NA    NA   
#> 10      2     20  1929        NA NA              NA      NA NA     NA    NA   
#> # … with 2,063,660 more rows, and 2 more variables: type3 <chr>,
#> #   ongoingrivalry <dbl>

# across_types defaults to 1
create_stateyears() %>% add_strategic_rivalries()

#> Joining, by = c("ccode", "year")

#> # A tibble: 16,731 x 7
#>    ccode statenme            year ideological interventionary positional spatial
#>    <dbl> <chr>              <int>       <dbl>           <dbl>      <dbl>   <dbl>
#>  1     2 United States of …  1816           0               0          1       1
#>  2     2 United States of …  1817           0               0          1       1
#>  3     2 United States of …  1818           0               0          1       1
#>  4     2 United States of …  1819           0               0          1       1
#>  5     2 United States of …  1820           0               0          1       0
#>  6     2 United States of …  1821           0               0          1       1
#>  7     2 United States of …  1822           0               0          1       1
#>  8     2 United States of …  1823           0               0          1       1
#>  9     2 United States of …  1824           0               0          1       1
#> 10     2 United States of …  1825           0               0          1       1
#> # … with 16,721 more rows

# }