Abbreviate CoW-MID Dispute-level Data (v. 5.0)

This is an abbreviated version of the dispute-level CoW-MID data.

cow_mid_disps

Format

A data frame with 2436 observations on the following 7 variables.

dispnum

a numeric vector for the CoW-MID dispute number

outcome

a numeric vector for the outcome of the MID

styear

a numeric vector for the start year of the MID

stmon

a numeric vector for the start month of the MID

settle

a numeric vector for the how dispute was settled

fatality

a numeric vector for the fatality level of the dispute

mindur

a numeric vector for the minimum duration of the MID

maxdur

a numeric vector for the maximum duration of the MID

hiact

a numeric vector for the highest action of the MID

hostlev

a numeric vector for the hostility level of the MID

recip

a numeric vector for whether a MID was reciprocated

Details

These data are purposely light on information; they're not intended to be used for dispute-level analyses, per se. They're intended to augment the directed dyadic dispute-year data by adding in variables that serve as exclusion rules to whittle the data from dyadic dispute-year to just dyad-year data.

References

Palmer, Glenn, and Roseanne W. McManus and Vito D'Orazio and Michael R. Kenwick and Mikaela Karstens and Chase Bloch and Nick Dietrich and Kayla Kahn and Kellan Ritter and Michael J. Soules. 2021. "The MID5 Dataset, 2011–2014: Procedures, coding rules, and description" Conflict Management and Peace Science.