Correlates of War Intra-State War Data (v. 4.1)

These are a modified version of the intra-state war data from the Correlates of War project. Data are version 4.1. The temporal domain is 1816-2007.

Usage

cow_war_intra

Format

A data frame with 1361 observations on the following 17 variables.

warnum

the Correlates of War war number

warname

the Correlates of War war name

wartype

a character vector for the type of war, either "local issues" or "central control"

year

a numeric vector for the year

cowintraonset

a dummy variable for whether this is a civil war onset (i.e. either the year in StartYear1 or StartYear2 in the raw data)

cowintraongoing

a numeric constant of 1

resume_combat

a dummy variable for whether this is a resumption of a conflict (i.e. StartYear2 is not -8)

primary_state

a dummy variable for whether the state is the primary state having the civil war

ccodea

the Correlates of War state code for the participant on Side A. -8 = not applicable (participant is not a state)

sidea

the name of the participant on Side A. -8 = not applicable (no additional party on this side)

ccodeb

the Correlates of War state code for the participant on Side B. -8 = not applicable (participant is not a state)

sideb

the name of the participant on Side B. -8 = not applicable (no additional party on this side)

intnl

a dummy variable for if this is an internationalized civil war

outcome

an unordered-categorical variable for the outcome of the civil war. Values include 1 (Side A wins), 2 (Side B wins), 3 (Compromise), 4 (war transformed into another type of war), 5 (war is ongoing at the end of 2007), 6 (stalemate), 7 (conflict continues below severity of war)

sideadeaths

the estimated deaths for the Side A participant (-9 = unknown, -8 = not applicable)

sidebdeaths

the estimated deaths for the Side B participant (-9 = unknown, -8 = not applicable)

ongo2007

a dummy variable for if this war is ongoing as of the end of 2007

Details

See data-raw directory for how these data were generated. In the Guinnea-Bissau Civil War (1998, 1999), the "Mane Junta" have the accented-e scrubbed to coincide with CRAN's character requirements.

References

Dixon, Jeffrey, and Meredith Sarkees. 2016. A Guide to Intra-State Wars: An Examination of Civil Wars, 1816-2014. Thousand Oaks, CA: Sage.

Sarkees, Meredith Reid, and Frank Wheldon Wayman. 2010. Resort to War: A Data Guide to Inter-State, Extra-State, Intra-State, and Non-State Wars, 1816-2007. Washington DC: CQ Press.