Download Some Extra Data for Peace Science Research

download_extdata() leverages R's inst directory flexibility to allow you to download some extra data and store it in the package.

download_extdata(overwrite = FALSE)

Arguments

overwrite
logical, defaults to FALSE. If FALSE, the function checks to see if you've already downloaded the data and, if you already have, it does nothing. If TRUE, the function redownloads the data.

Value

download_extdata() downloads some extra data stored on my website (http://svmiller.com) and sticks them in the extdata directory in the package. Right now, these data are just the directed dyad-year Correlates of War trade data.

A Description of Various Data Sets This Will Download

Running download_extdata() returns the following data that will be stored in the package's extdata directory.

Correlates of War Dyadic Trade Data Set (v. 4.0)

These are directed dyad-year-level data for national trade from the Correlates of War project.

COLUMN	DESCRIPTION

ccode1	a numeric vector for the Correlates of War state code for the first state

ccode2	a numeric vector for the Correlates of War state code for the second state

year	the year

flow1	imports of ccode1 from ccode2, in current million USD

flow2	imports of ccode2 from ccode1, in current million USD

smoothflow1	smoothed flow1 values

smoothflow2	smoothed flow2 values

References

Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009. "TRADING DATA: Evaluating our Assumptions and Coding Rules." Conflict Management and Peace Science. 26(5): 471-491.

Author

Steven V. Miller

Examples

if (FALSE) {
download_extdata()
}