False Gleditsch-Ward Directed Dyad-Years

This is a simple data set that communicates directed dyads in the Gleditsch-Ward data that appear in the same year, but not in any particular day in the year. They are used in an anti-join in the create_dyadyears() function in this package.

Usage

false_gw_dyads

Format

A data frame the following four variables.

gwcode1

a numeric vector for the Gleditsch-Ward state code for the first state

gwcode2

a numeric vector for the Gleditsch-Ward state code for the second state

year

a numeric vector for the year

in_ps

a constant that equals 1 if these data would appear in create_dyadyears() if you were not careful to remove them.

Details

Think of the directed Suriname and Republic of Vietnam dyad here as illustrative here. The Republic of Vietnam exits the Correlates of War state system on April 30, 1975 whereas Suriname enters the state system on November 25, 1975. Both appear in the same year, but not at the same time.