Directed dispute-year data (Gibler, Miller, and Little, 2016)

These are directed dispute-year data from the most recent version (2.1.1) of the Gibler-Miller-Little (GML) militarized interstate dispute (MID) data. They are used internally for merging into full dyad-year data frames.

gml_dirdisp

Format

A data frame with 10330 observations on the following 39 variables.

dispnum

the dispute number

ccode1

a numeric vector for the Correlates of War state code for the first state

ccode2

a numeric vector for the Correlates of War state code for the second state

year

a numeric vector for the year

midongoing

a constant of 1 for ongoing disputes

midonset

a numeric vector that equals 1 for the onset year of a given dispute

sidea1

is the first state (in ccode1) on the side that took the first militarized action?

sidea2

is the second state (in ccode2) on the side that took the first militarized action?

revstate1

is the first state (in ccode1) a revisionist state in the dispute?

revstate2

is the second state (in ccode2) a revisionist state in the dispute?

revtype11

what is the revtype1 value for ccode1?

revtype12

what is the revtype1 value for ccode2?

revtype21

what is the revtype2 value for ccode1?

revtype22

what is the revtype2 value for ccode2?

fatality1

what is the fatality value for ccode1?

fatality2

what is the fatality value for ccode2?

fatalpre1

what is the fatalpre value for ccode1?

fatalpre2

what is the fatalpre value for ccode2?

hiact1

what is the hiact value for ccode1?

hiact2

what is the hiact value for ccode2?

hostlev1

what is the hostlev value for ccode1?

hostlev2

what is the hostlev value for ccode2?

orig1

is ccode1 an originator (1) of the dispute or a joiner (0)?

orig2

is ccode2 an originator (1) of the dispute or a joiner (0)?

hiact

the highest level of action observed in the dispute

hostlev

the hostility level of action observed in the dispute

mindur

the minimum length of the dispute (in days)

maxdur

the maximum length of the dispute (in days)

outcome

the dispute-level outcome

settle

the settlement value for the dispute

fatality

the ordinal fatality level for the dispute

fatalpre

the fatalities (with precision, if known) for the dispute

stmon

the start month of the dispute (dispute-level)

endmon

the end month of the dispute (dispute-level)

recip

was the dispute reciprocated (i.e. did Side B also have a militarized action)?

numa

the number of participants on Side A

numb

the number of participants on Side B

ongo2010

was the dispute ongoing as of 2010?

version

a version identifier

Details

Data are the directed dispute-year data made available in version 2.1.1 of the GML MID data.

I would caution against using the revtype variables. They are not informative. They are however included for legacy reasons.

References

Gibler, Douglas M., Steven V. Miller, and Erin K. Little. 2016. “An Analysis of the Militarized Interstate Dispute (MID) Dataset, 1816-2001.” International Studies Quarterly 60(4): 719-730.