UCDP Armed Conflict Data (ACD) (v. 20.1)

These are (kind of) dyadic, but mostly state-level data, used internally for doing stuff with the UCDP armed conflict data

ucdp_acd

Format

A data frame with 4164 observations on the following 15 variables.

conflict_id

a conflict identifier, not to be confused with an episode identifier (which I don't think UCDP offers)

year

a numeric vector for the year

gwno_a

the Gleditsch-Ward state code for the state on side A of the armed conflict

gwno_a_2nd

the Gleditsch-Ward state code for the state that actively supported side A of the armed conflict with the use of troops

gwno_b

the Gleditsch-Ward state code for the actor on side B of the armed conflict

gwno_b_2nd

the Gleditsch-Ward state code for the state that actively supported side B of the armed conflict with the use of troops

incompatibility

a character vector for the main conflict issue ("territory", "government", "both")

intensity_level

a numeric vector for the intensity level in the calendar year (1 = minor (25-999 deaths), 2 = war (>1,000 deaths))

type_of_conflict

a character vector for the type of conflict ("extrasystemic", "interstate", "intrastate", "II"). "II" is a simple abbreviation of "internationalized intrastate"

start_date

a date of the first battle-related death in the conflict, not to be confused with the first battle-related death of the episode

start_prec

the level of precision for start_date

start_date2

a date of the first battle-related death in the episode, not to be confused with the first battle-related death of the conflict

start_prec2

the level of precision for start_date2

ep_end

a dummy variable for whether the conflict episode ended in the calendar year of observation

ep_end_date

the episode end date, if applicable

Details

The data-raw directory on the project's Github will show how I processed the multiple strings for when there are multiple states on a given side.

References

Gleditsch, Nils Petter; Peter Wallensteen, Mikael Eriksson, Margareta Sollenberg & Håvard Strand (2002) Armed Conflict 1946–2001: A New Dataset. Journal of Peace Research 39(5): 615–637.

Pettersson, Therese; Stina Högbladh & Magnus Öberg (2019). Organized violence, 1989-2018 and peace agreements. Journal of Peace Research 56(4): 589-603.